perf(cache): isolate handler and request state
