+++ packages/vinext/src/server/app-page-render-observation.ts
@@ -1,6 +1,6 @@
 import type { ReactNode } from "react";
 import { fnv1a64 } from "../utils/hash.js";
-import { _consumeRequestScopedCacheLife } from "vinext/shims/cache";
+import { _consumeRequestScopedCacheLife } from "vinext/shims/cache-request-state";
 import { consumeDynamicFetchObservations } from "vinext/shims/fetch-cache";
 import {
   consumeDynamicUsage,
