+++ packages/vinext/src/server/seed-cache.ts
@@ -31,7 +31,7 @@
 
 import fs from "node:fs";
 import path from "node:path";
-import type { CachedAppPageValue } from "vinext/shims/cache";
+import type { CachedAppPageValue } from "vinext/shims/cache-handler";
 import { isrCacheKey, isrSetPrerenderedAppPage } from "./isr-cache.js";
 import { buildAppPageCacheTags } from "./app-page-cache.js";
 import { getOutputPath, getRscOutputPath } from "../utils/prerender-output-paths.js";
