+++ packages/vinext/src/shims/request-state-types.ts
@@ -2,7 +2,7 @@ export type { HeadersAccessPhase, HeadersContext, VinextHeadersShimState } from 
 export type { I18nState } from "./i18n-state.js";
 export type { NavigationContext } from "./navigation.js";
 export type { NavigationState } from "./navigation-state.js";
-export type { CacheLifeConfig, CacheState } from "./cache.js";
+export type { CacheLifeConfig, CacheState } from "./cache-request-state.js";
 export type { PrivateCacheState } from "./cache-runtime.js";
 export type { FetchCacheState } from "./fetch-cache.js";
 export type { ExecutionContextLike } from "./request-context.js";
