+++ tests/app-router-isr-codegen.test.ts
@@ -38,6 +38,16 @@ describe("generateRscEntry ISR code generation", () => {
     expect(code).toContain("export default __createAppRscHandler({");
   });
 
+  it("configures the cache through the lightweight handler runtime", () => {
+    const code = generateRscEntry("/tmp/test/app", minimalRoutes);
+    expect(code).toContain(
+      'configureMemoryCacheHandler as __configureMemoryCacheHandler } from "vinext/shims/cache-handler"',
+    );
+    expect(code).not.toContain(
+      'configureMemoryCacheHandler as __configureMemoryCacheHandler } from "next/cache"',
+    );
+  });
+
   it("generated code stores root layout params separately from leaf params", () => {
     const routes = [
       {
